Convert publish script to TypeScript
