+++ vuu-ui/scripts/utils.ts
@@ -3,7 +3,7 @@ import { promisify } from "node:util";
 
 const exec = promisify(execCallback);
 
-export const execWait = async (command: string, cwd: string) => {
+export const execWait = async (command: string, cwd = ".") => {
   const { stdout, stderr } = await exec(command, {
     cwd,
   });
